replace the Socrata class with Sodapy
Disambiguates with the socrata-py package.
