+++ sodapy/__init__.py
@@ -1,6 +1,8 @@
 from sodapy.socrata import Socrata
+from sodapy.sodapy import Sodapy
 
 __all__ = [
     "Socrata",
+    "Sodapy",
 ]
 __version__ = "2.2.0"
